Fix zlib compression and decompression to work correctly on 64-bit operating systems

+++ core/PRP/NetMessage/plNetMsgStream.h
@@ -31,9 +31,9 @@ public:
     };
 
 private:
-    unsigned int fUncompressedSize;
+    size_t fUncompressedSize;
     unsigned char fCompressionType;
-    unsigned int fStreamLength;
+    size_t fStreamLength;
     unsigned char* fStream;
 
 public:
