+++ core/Util/plZlib.h
@@ -19,10 +19,12 @@
 
 #include "PlasmaDefs.h"
 
+#include <cstddef> // for size_t
+
 DllClass plZlib {
 public:
-    static bool Uncompress(unsigned char* bufOut, unsigned int& bufLenOut, const unsigned char* bufIn, unsigned int bufLenIn);
-    static bool Compress(unsigned char*& bufOut, unsigned int& bufLenOut, const unsigned char* bufIn, unsigned int bufLenIn);
+    static bool Uncompress(unsigned char* bufOut, size_t& bufLenOut, const unsigned char* bufIn, size_t bufLenIn);
+    static bool Compress(unsigned char*& bufOut, size_t& bufLenOut, const unsigned char* bufIn, size_t bufLenIn);
 };
 
 #endif
